+++ diesel/src/mysql/types/numeric.rs
@@ -2,20 +2,15 @@
 pub mod bigdecimal {
     extern crate bigdecimal;
 
-    use std::error::Error;
+    use self::bigdecimal::BigDecimal;
     use std::io::prelude::*;
 
     use mysql::Mysql;
-
-    use self::bigdecimal::BigDecimal;
-
     use types::{self, FromSql, IsNull, ToSql, ToSqlOutput};
+    use {deserialize, serialize};
 
     impl ToSql<types::Numeric, Mysql> for BigDecimal {
-        fn to_sql<W: Write>(
-            &self,
-            out: &mut ToSqlOutput<W, Mysql>,
-        ) -> Result<IsNull, Box<Error + Send + Sync>> {
+        fn to_sql<W: Write>(&self, out: &mut ToSqlOutput<W, Mysql>) -> serialize::Result {
             write!(out, "{}", *self)
                 .map(|_| IsNull::No)
                 .map_err(|e| e.into())
@@ -23,7 +18,7 @@ pub mod bigdecimal {
     }
 
     impl FromSql<types::Numeric, Mysql> for BigDecimal {
-        fn from_sql(bytes: Option<&[u8]>) -> Result<Self, Box<Error + Send + Sync>> {
+        fn from_sql(bytes: Option<&[u8]>) -> deserialize::Result<Self> {
             let bytes = not_none!(bytes);
             BigDecimal::parse_bytes(bytes, 10)
                 .ok_or_else(|| Box::from(format!("{:?} is not valid decimal number ", bytes)))